dynamic pydantic version validation
